Use component in page directly
